requirements.txt
PyYAML
flask
requests
flask-restful
ruamel.yaml
prometheus_flask_exporter
python-ldap
pyJWT
hvac==0.10.7
grpcio
grpcio-tools
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"